refactor : MemoryUserRepositoryTest 단위 테스트 코드 파일 패키지 변경
★ 참고사항 ★
커밋 변경 사항에는 기존 파일이 삭제되고 새로운 파일이 생성된다고 나오지만 실제로는 패키지 변경 사항 외에는 아무 수정 없습니다.
